Add some more tests to InvokeDroppedProcedureThread.
We add some tests which call UDFs.  There is a conjecture that
the mechanism mapping UDF names to HSQL metadata is corrupted by
database operations such as joining, pausing, restarting etc.
This tries to validate this conjecture.

We add 9 new UDFs, all of which have the same signature but
different implementations.  If the mapping is incorrect we will
call the wrong Java function.  By inspecting the answers we will
be able to tell that something is amiss.